небольшие доработки

+++ app/src/main/java/com/sofiamarchinskaya/hw1/models/framework/NoteRepository.kt
@@ -14,5 +14,5 @@ interface NoteRepository {
 
     fun insertCloud(note: Note)
 
-    suspend fun getLast():Long
+    suspend fun getLast(): Int
 }
